Move shell stuff out into its own file
